requirements.txt
astroid==2.3.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.4.4
requests==2.22.0
six==1.14.0
tqdm==4.42.0
urllib3==1.25.8
wrapt==1.11.2
python-bidi==0.4.2
lxml~=4.5.2